EY-1977: Flyttar den generiske delen av koden ut til libs så vi kan bruke samme mekanisme genereltˆ
