+++ apps/etterlatte-vilkaarsvurdering-kafka/src/main/kotlin/vilkaarsvurdering/services/VilkaarsvurderingService.kt
@@ -1,13 +1,10 @@
 package no.nav.etterlatte.vilkaarsvurdering.services
 
 import io.ktor.client.HttpClient
-import io.ktor.client.request.post
 import io.ktor.client.statement.HttpResponse
-import io.ktor.http.ContentType
-import io.ktor.http.contentType
 import kotlinx.coroutines.runBlocking
+import no.nav.etterlatte.libs.ktor.post
 import no.nav.etterlatte.vilkaarsvurdering.OpprettVilkaarsvurderingFraBehandling
-import vilkaarsvurdering.Invocation
 import vilkaarsvurdering.VilkaarsvurderingAPI
 import java.util.*
 
@@ -30,9 +27,3 @@ class VilkaarsvurderingServiceImpl(private val vilkaarsvurderingKlient: HttpClie
         vilkaarsvurderingKlient.post(VilkaarsvurderingAPI.opprett(behandlingId.toString()).medBody(url) { })
     }
 }
-
-suspend fun <T> HttpClient.post(invocation: Invocation<T>) =
-    this.post(invocation.url) {
-        contentType(ContentType.Application.Json)
-        invocation.body
-    }
