+++ libs/etterlatte-vilkaarsvurdering-model/src/main/kotlin/vilkaarsvurdering/VilkaarsvurderingAPI.kt
@@ -1,20 +1,12 @@
 package vilkaarsvurdering
 
+import no.nav.etterlatte.libs.common.Route
 import no.nav.etterlatte.vilkaarsvurdering.OpprettVilkaarsvurderingFraBehandling
 
 object VilkaarsvurderingAPI {
     const val basisrute = "api/vilkaarsvurdering"
     fun kopier(behandlingsId: String) =
-        Route("/$behandlingsId/kopier", OpprettVilkaarsvurderingFraBehandling::class.java)
+        Route("/$behandlingsId/kopier", basisrute, OpprettVilkaarsvurderingFraBehandling::class.java)
 
-    fun opprett(behandlingsId: String) = Route("/$behandlingsId/opprett", Unit::class.java)
+    fun opprett(behandlingsId: String) = Route("/$behandlingsId/opprett", basisrute, Unit::class.java)
 }
-
-data class Route<T>(
-    val url: String,
-    val clazz: Class<T>
-) {
-    fun medBody(baseURL: String, f: () -> T) = Invocation("$baseURL/$VilkaarsvurderingAPI.basisrute/$url", f())
-}
-
-data class Invocation<T>(val url: String, val body: T)
